feat: adding feature claim and unclaim

+++ controllers/ClaimController.js
@@ -3,21 +3,32 @@ import VoucherClaim from "../models/VoucherClaimModel.js";
 import { responseSuccess } from "../utils/responseTemplate.js";
 import Voucher from "../models/VoucherModel.js";
 
-export const claimVoucerController = async (req, res, next) => {
+export const claimVoucer = async (req, res, next) => {
   try {
     const { id } = req.body;
-
     const id_user = req.user.id;
 
     if (!id) {
-      throw new CustomError("id voucher must be fill", 400);
+      throw new CustomError("ID voucher must be filled", 400);
     }
 
-    const claimedVoucher = await VoucherClaim.create({
+    const voucher = await Voucher.findByPk(id);
+
+    if (!voucher) {
+      throw new CustomError("Voucher not found", 404);
+    }
+
+    if (voucher.status === true) {
+      throw new CustomError("Voucher already claimed", 400);
+    }
+
+    const data = {
       id_voucher: id,
       id_user,
       tanggal_claim: new Date(),
-    });
+    };
+
+    const claimedVoucher = await VoucherClaim.create(data);
 
     await Voucher.update(
       { status: true },
@@ -28,26 +39,31 @@ export const claimVoucerController = async (req, res, next) => {
       }
     );
 
-    if (!claimedVoucher) {
-      throw new CustomError("Failed claim voucher", 400);
-    }
-
-    responseSuccess(res, 201, "Success claimed voucher", [claimedVoucher]);
+    responseSuccess(res, 201, "Successfully claimed voucher", [claimedVoucher]);
   } catch (err) {
     next(err);
   }
 };
 
-export const unclaimVoucherController = async (req, res, next) => {
-  const { id } = req.query;
+export const unclaimVoucher = async (req, res, next) => {
+  const { id } = req.params;
+  const id_user = req.user.id;
+
   try {
     if (!id) {
-      throw new CustomError("id voucher must be fill", 400);
+      throw new CustomError("ID voucher must be filled", 400);
     }
 
-    const unclaimedVoucher = await VoucherClaim.destroy({
+    const voucherClaim = await VoucherClaim.findByPk(id);
+
+    if (!voucherClaim) {
+      throw new CustomError("Claim not found", 404);
+    }
+
+    await VoucherClaim.destroy({
       where: {
         id: id,
+        id_user,
       },
     });
 
@@ -60,11 +76,7 @@ export const unclaimVoucherController = async (req, res, next) => {
       }
     );
 
-    if (!unclaimedVoucher) {
-      throw new CustomError("Failed remove claimed voucher", 400);
-    }
-
-    responseSuccess(res, 200, "Success unclaim voucher");
+    responseSuccess(res, 200, "Successfully unclaimed voucher");
   } catch (err) {
     next(err);
   }
